+++ src/types.ts
@@ -1,14 +1,12 @@
-export const FILES = ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h'] as const;
+export const FILE_NAMES = ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h'] as const;
+export type FileName = typeof FILE_NAMES[number];
 
-export type File = typeof FILES[number];
-
-export const RANKS = ['1', '2', '3', '4', '5', '6', '7', '8'] as const;
-
-export type Rank = typeof RANKS[number];
+export const RANK_NAMES = ['1', '2', '3', '4', '5', '6', '7', '8'] as const;
+export type RankName = typeof RANK_NAMES[number];
 
 export type Square = number;
 
-export type SquareName = `${File}${Rank}`;
+export type SquareName = `${FileName}${RankName}`;
 
 export type BySquare<T> = T[];
 
@@ -21,7 +19,6 @@ export type ByColor<T> = {
 };
 
 export const ROLES = ['pawn', 'knight', 'bishop', 'rook', 'queen', 'king'] as const;
-
 export type Role = typeof ROLES[number];
 
 export type ByRole<T> = {
@@ -29,7 +26,6 @@ export type ByRole<T> = {
 };
 
 export const CASTLING_SIDES = ['a', 'h'] as const;
-
 export type CastlingSide = typeof CASTLING_SIDES[number];
 
 export type ByCastlingSide<T> = {
@@ -64,7 +60,6 @@ export function isNormal(v: Move): v is NormalMove {
 }
 
 export const RULES = ['chess', 'antichess', 'kingofthehill', '3check', 'atomic', 'horde', 'racingkings', 'crazyhouse'] as const;
-
 export type Rules = typeof RULES[number];
 
 export interface Outcome {